moved Icons.tsx to /icons; moved mainCats, moreCats, routeMap, IconsMap to new /constants folder; styling changes
