+++ src/components/MobileNav.tsx
@@ -11,12 +11,12 @@ const MobileNav = () => {
       <div className="block md:hidden">
         <button
           onClick={() => setIsOpen(!isOpen)}
-          className={`text-black-500 hover:text-black-400 flex items-center rounded px-3 py-2 ${
+          className={`text-black-500 hover:text-black-400 flex items-center rounded px-3 py-1 ${
             isOpen ? 'hidden' : 'block'
           }`}
         >
           <svg
-            className={`h-6 w-6 fill-current ${isOpen ? 'hidden' : 'block'}`}
+            className={`${isOpen ? 'hidden' : 'block'} h-7 w-7 fill-black`}
             viewBox="0 0 20 20"
             xmlns="http://www.w3.org/2000/svg"
           >
