+++ src/components/SidebarItems.tsx
@@ -1,47 +1,10 @@
 'use client';
 import { type FC, useState } from 'react';
-import {
-  AboutIcon,
-  EventsIcon,
-  FeedbackIcon,
-  HistoryIcon,
-  HomeIcon,
-  type IconType,
-  LikedIcon,
-  RightChevron,
-  SettingsIcon,
-} from './Icons';
 import { useRouter, usePathname } from 'next/navigation';
+import { IconMap, type allCats, routeMap } from '@src/constants/categories';
+import { RightChevron } from '../icons/Icons';
 
-const mainCats = ['Home', 'History', 'Liked', 'Events'] as const;
-const moreCats = ['Settings', 'About', 'Feedback'] as const;
-
-type union = (typeof mainCats)[number] | (typeof moreCats)[number];
-const IconMap: {
-  [key in union[number]]: IconType;
-} = {
-  Home: HomeIcon,
-  History: HistoryIcon,
-  Liked: LikedIcon,
-  Events: EventsIcon,
-  Settings: SettingsIcon,
-  About: AboutIcon,
-  Feedback: FeedbackIcon,
-};
-
-const routeMap: {
-  [key in union[number]]: string;
-} = {
-  Home: '/',
-  History: '/history',
-  Liked: '/liked',
-  Events: '/events',
-  Settings: '/settings',
-  About: '/about',
-  Feedback: '/feedback',
-};
-
-const SidebarItems: FC<{ cat: union[number] }> = ({ cat }) => {
+const SidebarItems: FC<{ cat: allCats[number] }> = ({ cat }) => {
   const Icon = IconMap[cat];
   const router = useRouter();
   const route = routeMap[cat];
